Remove 'dbDisconnect', 'dbReconnect' generics and methods
These were needed for the 'DB1' format but are no longer needed
since I changed the way the connections are handled.  Originally,
dbDisconnect was needed to close the 'file' connection in 'DB1'
but since this is now just opened and closed within each operation,
there is no need for a separate function to close the connect.

Hopefully, this will allow interoperation with 'filehashSQLite'.