Trim trailing '/' in initializeRDS
Got the following report:

I just want to report you a minor problem that I've noticed with dbInit.

> dbInit("C:/Logs", type="RDS")
'filehashRDS' database 'Logs'
> dbInit("C:/Logs/", type="RDS")
Error in validObject(.Object) :
  invalid class "filehashRDS" object: directory 'C:\Logs\' does not exist
>

This appears to be a Windows system and the problem is that 'file.exists'
appears to be returning FALSE if the directory has a trailing '/'.